Removing a Visual Studio 2013 warning in tokenize.cpp
1>tokenize.cpp(3218): warning C4244: '=' : conversion from
'MathLib::bigint' to 'size_t', possible loss of data

+++ lib/tokenize.cpp
@@ -3215,7 +3215,7 @@ bool Tokenizer::simplifySizeof()
                 if (Token::Match(decltok,"%name% [") && Token::simpleMatch(decltok->linkAt(1), "] [")) {
                     const Token *tok2 = decltok->linkAt(1);
                     while (Token::Match(tok2, "] [ %num% ]")) {
-                        sz = sz * MathLib::toLongNumber(tok2->strAt(2));
+						sz = sz * static_cast<unsigned long>( MathLib::toLongNumber(tok2->strAt(2)) );
                         tok2 = tok2->linkAt(3);
                     }
                     if (Token::simpleMatch(tok2, "] ["))
